Niina and Timothy's changes
We added the unstandardized/standardized tables of the training data set. Also, we included the graphs that have alcohol on the y-axis and the other potential predicting variables on x-axis. It might be best to use these graphs to decide which variables we want to use as predicting variables (alcohol and something). We also commented the workflow until the predictor variables are decided upon (please do not edit the workflow).